requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pandas>=2.0.0
pyproj>=3.6.0
